@@ -17,29 +17,41 @@
             "homepage": "http://mwop.net/"
         }
     ],
+    "extra": {
+        "branch-alias": {
+            "dev-master": "1.1-dev",
+            "dev-develop": "1.2-dev"
+        }
+    },
     "require": {
-        "php": ">=5.3.23",
+        "php": ">=5.5",
         "ext-mongo": "*",
-        "zendframework/zend-paginator": "2.*",
-        "zendframework/zend-servicemanager": "2.*",
-        "zendframework/zend-stdlib": "2.*"
+        "zendframework/zend-i18n": "~2.0",
+        "zendframework/zend-paginator": "~2.0",
+        "zendframework/zend-servicemanager": "~2.0",
+        "zendframework/zend-stdlib": "~2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~3.7",
-        "zendframework/zend-form": "2.*",
-        "zendframework/zend-loader": "2.*",
-        "zendframework/zend-modulemanager": "2.*",
-        "zendframework/zend-mvc": "2.*",
-        "zendframework/zend-serializer": "2.*",
-        "zendframework/zend-view": "2.*"
+        "phpunit/phpunit": "~4.7",
+        "squizlabs/php_codesniffer": "^2.3.1",
+        "zendframework/zend-form": "~2.0",
+        "zendframework/zend-log": "~2.0",
+        "zendframework/zend-modulemanager": "~2.0",
+        "zendframework/zend-mvc": "~2.0",
+        "zendframework/zend-serializer": "~2.0",
+        "zendframework/zend-view": "~2.0"
     },
     "autoload": {
-        "psr-0": {
-            "PhlyMongo": "src/",
-            "PhlyMongoTest": "test/"
+        "psr-4": {
+            "PhlyMongo\\": "src/"
         },
         "classmap": [
             "Module.php"
         ]
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "PhlyMongoTest\\": "test/"
+        }
     }
 }
