@@ -1,4 +1,8 @@
 <?php
+/**
+ * @license   http://opensource.org/licenses/BSD-2-Clause BSD-2-Clause
+ * @copyright Copyright (c) 2014 Matthew Weier O'Phinney
+ */
 
 namespace PhlyMongo;
 
@@ -8,7 +12,7 @@ class Module
     {
         return array('Zend\Loader\StandardAutoloader' => array(
             'namespaces' => array(
-                __NAMESPACE__ => __DIR__ . '/src/' . __NAMESPACE__,
+                __NAMESPACE__ => __DIR__ . '/src/',
             ),
         ));
     }
