@@ -1,10 +1,10 @@
 <?php
+/**
+ * @license   http://opensource.org/licenses/BSD-2-Clause BSD-2-Clause
+ * @copyright Copyright (c) 2014 Matthew Weier O'Phinney
+ */
 
 namespace PhlyMongo;
-
-use InvalidArgumentException;
-use MongoCursor;
-use Zend\Stdlib\Hydrator\HydratorInterface;
 
 class HydratingPaginatorAdapter extends PaginatorAdapter
 {
